refactor: merge side.py and calls.py to funcs.py
