+++ src/snakia/utils/__init__.py
@@ -1,10 +1,9 @@
 from .attrs import get_attrs, get_or_set_attr
-from .calls import call, caller
+from .funcs import call, caller, side, side_func
 from .exceptions import catch, throw
 from .frames import frame
 from .gil import GIL_ENABLED, nolock
 from .inherit import inherit
-from .side import side, side_func
 from .this import this
 from .to_async import to_async
