+++ src/snakia/utils/funcs.py
@@ -1,4 +1,4 @@
-from typing import Callable, ParamSpec, TypeVar
+from typing import Any, Callable, ParamSpec, TypeVar
 
 P = ParamSpec("P")
 T = TypeVar("T")
@@ -10,3 +10,11 @@ def call(f: Callable[P, T], *args: P.args, **kwargs: P.kwargs) -> T:
 
 def caller(f: Callable[P, T], *args: P.args, **kwargs: P.kwargs) -> Callable[..., T]:
     return lambda *_, **__: f(*args, **kwargs)
+
+
+def side(value: T, *_: Any, **__: Any) -> T:
+    return value
+
+
+def side_func(value: T, *_: Any, **__: Any) -> Callable[..., T]:
+    return lambda *_, **__: value
